backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Bartosz Woźniak <bartoszwoz0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.1.1"
django-environ = "0.9.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.0"
django-extensions = "3.2.1"
requests = "2.28.1"

[tool.poetry.dev-dependencies]
pytest = "7.1.3"
pytest-django = "4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
